Introduce Deconstructor <: MemberPattern <: Executable
