+++ src/java.base/share/classes/jdk/internal/access/JavaLangReflectAccess.java
@@ -106,5 +106,5 @@ public interface JavaLangReflectAccess {
     public <T> T newInstance(Constructor<T> ctor, Object[] args, Class<?> caller)
         throws IllegalAccessException, InstantiationException, InvocationTargetException;
 
-    public String getMangledName(Deconstructor<?> d);
+    public String getMangledName(MemberPattern<?> d);
 }
